feat(feature/feature-security): add api review, aou follow , and api report summary

+++ src/main/java/com/quocchung/cntt1/techcycle_system/dtos/response/Reaction/ReactionUserInfo.java
@@ -16,5 +16,5 @@ public class ReactionUserInfo {
     private String ward;
     private String province;
     private String avatarUrl;
-    private Integer trustScore;
+    private Double trustScore;
 }
